### UPDATES AND ADDED FEATURES
- Created an email column in the owners table.
- Add values to visits table and email column.
- Created indexes.
-  Added three performance query tests.


### Screenshots
**First Query Before and after creating index**
![performance ss1 before](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/e240ed24-33cc-4c7f-a0a3-123927aaae5f)
*From 1058ms TO 269ms*
![performance ss1 after](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/74c0d47c-929e-42b1-92ee-4473db313972)

**Second Query Before and after creating index**
![performance ss2 before](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/99407e89-e038-4f60-b86d-902ec70d2a00)
*From 3958ms TO 1748ms*
![performance ss2 after](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/0ef4a3b8-edb4-4243-87be-0880bcf27bef)

**Third Query Before and after creating index**
![performance ss3 before](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/31e58f35-2d2e-42f8-a5c0-7abd764c185d)
*From 2345ms TO 0.08ms*
![performance ss3 after](https://github.com/ClintonYounge/Vet-Clinic-DB/assets/118696200/bc5ba477-299b-4868-8f6e-306af8742f1f)